+++ DataManager/ReportForm/src/@app.ts
@@ -2,11 +2,6 @@ import { ReportForm, DoGet } from "./Product/ReportForm/ReportForm";
 
 /* eslint @typescript-eslint/no-unused-vars: off */
 
-const VERSION = "0.4.4";
-export function getAppVersion(): string {
-    return VERSION;
-}
-
 function doGet(e: DoGet) {
     return ReportForm.doGet(e);
 }
