+++ DataManager/ReportForm/src/Product/ReportForm/Layer4/LINEPostCommands/VersionGetLINEPostCommand.ts
@@ -1,7 +1,7 @@
-import { getAppVersion } from "../../../../@app";
+import { Version } from "../../Version";
 import { LINEPostCommand } from "./@LINEPostCommand";
 export class VersionGetLINEPostCommand extends LINEPostCommand {
     public invoke(): void {
-        this.replyMessage(this.event.replyToken, [getAppVersion()]);
+        this.replyMessage(this.event.replyToken, [Version.toolVersion]);
     }
 }
