+++ DataManager/ReportForm/src/Product/ReportForm/Version.ts
@@ -1,0 +1,6 @@
+export class Version {
+    private static readonly _toolVersion = '0.4.4';
+    public static get toolVersion() {
+        return this._toolVersion;
+    }
+}
